requirements.txt
# Core ML
torch==2.2.2
transformers==4.40.2
sentence-transformers==2.7.0      # 升级，兼容新 huggingface_hub
huggingface-hub==0.34.4           # 与 transformers / tokenizers 一致
tokenizers==0.19.1                # transformers 4.40.2 依赖

# Vector DB
faiss-cpu==1.7.4

# Data
pandas==2.2.2
numpy==1.26.4

# API server
fastapi==0.110.0
uvicorn==0.29.0

# Visualization / optional
matplotlib==3.8.4
umap-learn==0.5.5
scikit-learn==1.4.2
